Domain outbox reviewed

+++ back/rbkApiModules/rbkApiModules.Commons.Core/Events/Events/Contracts/EventNameAttribute.cs
@@ -3,7 +3,7 @@ namespace rbkApiModules.Commons.Core;
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
 public sealed class EventNameAttribute : Attribute
 {
-    public EventNameAttribute(string name, int version = 1)
+    public EventNameAttribute(string name, short version = 1)
     {
         Name = name ?? throw new ArgumentNullException(nameof(name));
         Version = version;
@@ -11,5 +11,5 @@ public sealed class EventNameAttribute : Attribute
 
     public string Name { get; }
 
-    public int Version { get; }
+    public short Version { get; }
 }
