+++ back/rbkApiModules/rbkApiModules.Commons.Core/Events/Events/Envelope/EventEnvelope.cs
@@ -7,7 +7,7 @@ public sealed class EventEnvelope<TEvent>
 {
     public Guid EventId { get; init; }
     public string Name { get; init; } = default!;
-    public int Version { get; init; }
+    public short Version { get; init; }
     public DateTime OccurredUtc { get; init; }
     public string TenantId { get; init; } = default!;
     public string Username { get; init; } = default!;
